Added package blacklist
The package blacklist allows you to stop brewpoll from checking packages
that used to be shipped in your channels, but which are now sourced from
an alternative channel. Since you probably don't want to be alerted that
your version is out of date, blacklisting the package makes brewpoll
much quieter.